src final

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double weightInKilo, double heightInMeters) {
+    public int calculate(double weightInKilo, double heightInMeters) {
         double bmiIndex = (weightInKilo / (heightInMeters * heightInMeters));
         return (int) bmiIndex;
     }
